+++ app/uk/gov/hmrc/helptosave/controllers/HelpToSaveAuth.scala
@@ -21,7 +21,7 @@ import uk.gov.hmrc.auth.core.AuthProvider.GovernmentGateway
 import uk.gov.hmrc.auth.core._
 import uk.gov.hmrc.auth.core.authorise.Predicate
 import uk.gov.hmrc.auth.core.retrieve.Retrievals
-import uk.gov.hmrc.helptosave.util.{Logging, NINO, toFuture}
+import uk.gov.hmrc.helptosave.util.{Logging, NINO, WithMdcExecutionContext, toFuture}
 import uk.gov.hmrc.play.bootstrap.controller.BaseController
 
 import scala.concurrent.Future
