+++ app/uk/gov/hmrc/helptosave/util/WithMdcExecutionContext.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.helptosave.controllers
+package uk.gov.hmrc.helptosave.util
 
 import uk.gov.hmrc.http.logging.LoggingDetails
 import uk.gov.hmrc.play.http.logging.MdcLoggingExecutionContext
